Allow visualization of tesla streaming data on a map
Simply call

node ./visualize.js --db <database bane> --replay <number of minutes back>

to start the server and then point a web browser to

http://localhost:8766

This will show a Google map with the position and some statistics on the car.
By giving the --replay argument you can go back in time and visualize the path
the car has taken. This runs at roughly 30x speedup. The map keeps updating in
real time as long as the server process is running.

The color of the track visualizes the speed driven. Slower speeds are shown as
shorter wavelength (with parked being a fairly deep blue). Faster speeds in
longer and longer wavelength with speeds above 65 being solid red.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>