Move HTML source into a separate file
Load the file and replace some MAGIC_ values with the correct data at runtime.
With this there's an easy to edit and maintain 'map.html' file that gets used
as the basis for the file served.

Also extend the window for searching for timestamps to ten minutes (as I found
two instances where there was a gap of more than five minutes in my data).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>