Add date and time picker to the energy UI
This is pretty rough still. Dates aren't initialized with the last value.
Hitting ESC while picking a date creates an invalid query string, lots of
other problems. But it mostly works.

This also adds the MIT licensed jquery and jquery-ui components so people
don't need to hunt and seek for them.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>