Deliver fewer data points to the browser in energy mode
Previously we delivered all available data, which could be many megabytes
and millions of values - even though the browser window has no way of
actually plotting all these. This made the response times slow and wasted
bandwidth.

With thie commit we summarize the values delivered in the server process
and only deliver up to 10000 values for speed and SOC and 20000 values for
energy (this way we show both maximum energy used and maximum regen).

This also fixes a bug where parsing of the to/from parameter was too
inflexible (when switching from stats to energy we don't get hh-mm-ss).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>