Change charge chart to line chart and move to energy page
This seems to make much more sense - the trip page has daily summary
information - the charge chart with minute granularity didn't fit in with that.

In the energy chart on the other hand this is all now nicely aligned and gives
a reasonably complete picture.

Also, displaying current and voltage as lines instead of bars is much more
intuitive. Added the hovering tooltip gives us the kW, too.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>