Switch visualize.js to using express
This seems to make it easier to write a maintainable web application. And it
helps with things like adding authentication, etc.

One immediate benefit is delivery of static files - we move all of them into a
subdirectory (otherfiles) and then use the express.static function to have them
all delivered to the client.

Other than that this is a 1:1 conversion without added functionality.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>