Fix real time tracking
If no data came in from the streaming API between to AJAX calls (2
seconds - this can happen when stopped at a traffic light) then
visualize.js would continue on into the future until the end time is
reached. That makes no sense at all, so instead make sure that the
"lastTime" is never later than the "currentTime".

This also fixes the detection of the fake JSON data to recognize the
possible end of the data series.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>