Add more functions to lib.js and similar functions to visualize.js
This is a fairly big commit, but hard to break into smaller pieces.

First it adds several more functions to lib.js and then uses those in
welcome, map, stats and energy pages. This makes the code more consistent,
easier to read and much more robust regarding query arguments (it now
deals with '-' as well as ' ' and '%20' as separators for date and time in
the query strings, for example).

This adds similar helper functions to visualize.js (it's rather awkward to
share the same helper functions between client side js and node.js on the
server side, so I opted to mostly duplicate the functions - but I also
changed the way they are used a little bit to make them more applicable
for the server side.

visualize.js now redirects the requests for the individual pages to new
request URLs that have all elements of the query arguments filled - this
allows simple consistent parsing everywhere.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>