Stack vampire loss ontop of energy used
This may not be what we want once 5.0 is widely used, but for now it seems
useful.

In order to do this I need to upgrade the side-by-side plugin to a forked
version that works with the stack plugin (and add the stack plugin).

The positioning of the value labels in the bars is quite a hack but seems to
work for me.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>