Fix calculations for charging
When charging at a Supercharger, 'charger_actual_current' is zero as the
internal charger isn't used. Instead we need to look at the battery_current!

Also, when at the Supercharger today to test this I learned the hard way that
with bad 3G connectivity there can be long patches with no valid chargeState
data - even while the car is charging. So let's not pull the charge voltage /
current down to zero just because samples are missing.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>