Change the way charging is calculated for daily summary
Getting the charging data from the streaming data sounds appealing because of
the high resolution of that data (250ms when driving, down to 120s when parked
and idle). But since energy is given as integer value there appears to be some
serious cumulative error. So it appears much more useful to get this from the
chargeState data.

Sadly that has only been recorded on a regular basis (every minute) since
commit e77fafdf2024 on Aug 28. So before that data we should still keep the old
calculations around (that part isn't implemented, yet).

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>