Improved estimates for energy used for the stats page
There are different error cases that all contribute to incorrect data. The
worst one is if the streaming service is down (or you are out of range) while
you are driving. Even when manually filling in some data points (to at least
not have all the energy used be accounted as Vampire drain), the old algorithm
would get things relentlessly and obviously wrong.

This does several things to improve the situation:

- sort the data returned from the streaming algorithm
  (this one pains me, but if you are forced to add entries to the database
  to compensate for lacking streaming data, the streaming data is no longer
  in chronologic order and the algorithm flat out fails)
- calculate the used energy two different ways:
-- once by the old "discrete integration over time" using the streaming data
-- once by observing the charge state while driving (as that gives us range in
   1/10 mile resolution) and converting it to power used
- for unexplained reasons sometimes charge state data is missing, so we need
  to do both calculations and figure out which one to trust.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>