Add minimal profiling for stats
This shows just how much time we spend on the database query and later
processing.

I was hoping to be able to skip the sort step if the data is already sorted and
found out that it frequently isn't, even if the user hasn't touched the
database. The data appears to arrive out of order from Tesla.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>