Use the conversion functions
It turns out this was not as straight forward as I assumed when implementing
things. Given the massive amount of data that might have to be converted the
lookup and function call overhead made using the more elegant conversion
function way too slow when convering the values in the large arrays. But at
least for these two values there was no difference so I used this "just
because".

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>