Move visualize.js authentication data into the config file
As Hans requested, the authentication data is now maintained in the config.json
file. This makes a ton more sense as it will no longer be overwritten every
time a user updates to a new version of this project.

Als, I went back to making authentication optional. If no user/password pairs
are found in the config file, visualize.js defaults to being completely open.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>