Safety first - use minutes, not milliseconds
For one, most people will think about this in terms of minutes. But more
importantly, it's way too easy to miscount the number of zeroes (or worse, to
think the number is in seconds) and then get blocked for making way too many
requests.

If people /really/ need higher precision, they can give the times as decimals.
So -N 2.5 will give you 150000ms (or 2 minutes 30 seconds). I'll still argue
that this is much better and much safer.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>