added function to set token
Bearer token can be set with set_token() function so it can be reused
between sessions and avoids use of login and password. For backward
compatibility, calling all(), vehicles(), or get_vid() still generates
a token for the life of the session. Both old tokens and new ones will
work until invalidated by changing your password on the Tesla website.